extract reporter logic to a class
